requirements.txt
psycopg2-binary
Flask
Flask-Login
yandex-pgmigrate
requests